+++ packages/extension-base/src/koni/background/handlers/Extension.ts
@@ -1122,16 +1122,18 @@ export default class KoniExtension {
     return this.getPrice();
   }
 
-  private getBalance (reset?: boolean): BalanceJson {
+  private async getBalance (reset?: boolean) {
     return this.#koniState.getBalance(reset);
   }
 
-  private subscribeBalance (id: string, port: chrome.runtime.Port): BalanceJson {
+  private async subscribeBalance (id: string, port: chrome.runtime.Port) {
     const cb = createSubscription<'pri(balance.getSubscription)'>(id, port);
 
     const balanceSubscription = this.#koniState.subscribeBalance().subscribe({
       next: (rs) => {
-        cb(rs);
+        const data = { details: rs } as BalanceJson;
+
+        cb(data);
       }
     });
 
@@ -1141,7 +1143,7 @@ export default class KoniExtension {
       this.cancelSubscription(id);
     });
 
-    return this.getBalance(true);
+    return await this.getBalance(true);
   }
 
   private getCrowdloan (reset?: boolean): CrowdloanJson {
@@ -3474,6 +3476,10 @@ export default class KoniExtension {
       return await this.#koniState.reloadNft();
     } else if (data === 'staking') {
       return await this.#koniState.reloadStaking();
+    } else if (data === 'balance') {
+      return await this.#koniState.reloadBalance();
+    } else if (data === 'crowdloan') {
+      return await this.#koniState.reloadCrowdloan();
     }
 
     return Promise.resolve(false);
@@ -3877,7 +3883,7 @@ export default class KoniExtension {
     // only works with current account, fix later
     const assetInfoMap = this.#koniState.getAssetRegistry();
     const chainInfoMap = this.#koniState.getChainInfoMap();
-    const balanceMap = this.#koniState.getBalance().details;
+    const balanceMap = (await this.#koniState.getBalance()).details;
     const substrateApiMap = this.#koniState.getSubstrateApiMap();
     const balanceService = this.#koniState.balanceService;
 
@@ -3905,7 +3911,7 @@ export default class KoniExtension {
       address,
       amount: data?.amount || '0',
       assetInfoMap: this.#koniState.getAssetRegistry(),
-      balanceMap: this.#koniState.getBalance().details,
+      balanceMap: (await this.#koniState.getBalance()).details,
       chainInfoMap: this.#koniState.getChainInfoMap(),
       poolInfo: yieldPoolInfo,
       substrateApiMap: this.#koniState.getSubstrateApiMap()
@@ -3932,7 +3938,7 @@ export default class KoniExtension {
         address,
         amount: data?.amount || '0',
         assetInfoMap: this.#koniState.getAssetRegistry(),
-        balanceMap: this.#koniState.getBalance().details,
+        balanceMap: (await this.#koniState.getBalance()).details,
         chainInfoMap: this.#koniState.getChainInfoMap(),
         poolInfo: yieldPoolInfo,
         substrateApiMap: this.#koniState.getSubstrateApiMap()
@@ -3968,7 +3974,7 @@ export default class KoniExtension {
         address,
         amount,
         assetInfoMap: this.#koniState.getAssetRegistry(),
-        balanceMap: this.#koniState.getBalance().details,
+        balanceMap: (await this.#koniState.getBalance()).details,
         chainInfoMap: this.#koniState.getChainInfoMap(),
         poolInfo: yieldPoolInfo,
         substrateApiMap: this.#koniState.getSubstrateApiMap()
@@ -4018,12 +4024,12 @@ export default class KoniExtension {
     return this.#koniState.getYieldPositionInfo();
   }
 
-  private handleValidateYieldProcess (inputData: ValidateYieldProcessParams) {
+  private async handleValidateYieldProcess (inputData: ValidateYieldProcessParams) {
     const params: OptimalYieldPathParams = {
       address: inputData.address,
       amount: inputData.amount || '0',
       assetInfoMap: this.#koniState.getAssetRegistry(),
-      balanceMap: this.#koniState.getBalance().details,
+      balanceMap: (await this.#koniState.getBalance()).details,
       chainInfoMap: this.#koniState.getChainInfoMap(),
       poolInfo: inputData.yieldPoolInfo,
       substrateApiMap: this.#koniState.getSubstrateApiMap()
@@ -4423,9 +4429,9 @@ export default class KoniExtension {
       case 'pri(price.getSubscription)':
         return await this.subscribePrice(id, port);
       case 'pri(balance.getBalance)':
-        return this.getBalance();
+        return await this.getBalance();
       case 'pri(balance.getSubscription)':
-        return this.subscribeBalance(id, port);
+        return await this.subscribeBalance(id, port);
       case 'pri(crowdloan.getCrowdloan)':
         return this.getCrowdloan();
       case 'pri(crowdloan.getCrowdloanContributions)':
@@ -4475,7 +4481,7 @@ export default class KoniExtension {
       case 'pri(yield.submitRedeem)':
         return await this.handleYieldRedeem(request as RequestYieldFastWithdrawal);
       case 'pri(yield.validateProcess)':
-        return this.handleValidateYieldProcess(request as ValidateYieldProcessParams);
+        return await this.handleValidateYieldProcess(request as ValidateYieldProcessParams);
       case 'pri(yield.staking.submitUnstaking)':
         return await this.yieldSubmitUnstaking(request as RequestUnbondingSubmit);
       case 'pri(yield.staking.submitWithdraw)':