+++ packages/extension-koni-ui/src/Popup/Home/Tokens/DetailModal.tsx
@@ -4,7 +4,7 @@
 import { APIItemState } from '@subwallet/extension-base/background/KoniTypes';
 import { BalanceItem } from '@subwallet/extension-base/types';
 import { isSameAddress } from '@subwallet/extension-base/utils';
-import { AccountTokenBalanceItem, RadioGroup } from '@subwallet/extension-koni-ui/components';
+import { AccountTokenBalanceItem, EmptyList, RadioGroup } from '@subwallet/extension-koni-ui/components';
 import { BaseModal } from '@subwallet/extension-koni-ui/components/Modal/BaseModal';
 import { ScreenContext } from '@subwallet/extension-koni-ui/contexts/ScreenContext';
 import { useSelector } from '@subwallet/extension-koni-ui/hooks';
@@ -12,9 +12,10 @@ import useTranslation from '@subwallet/extension-koni-ui/hooks/common/useTransla
 import { ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { TokenBalanceItemType } from '@subwallet/extension-koni-ui/types/balance';
 import { isAccountAll } from '@subwallet/extension-koni-ui/utils';
-import { Form, ModalContext, Number } from '@subwallet/react-ui';
+import { Form, Icon, ModalContext, Number } from '@subwallet/react-ui';
 import BigN from 'bignumber.js';
 import CN from 'classnames';
+import { ArrowCircleLeft, Coins } from 'phosphor-react';
 import React, { useContext, useEffect, useMemo } from 'react';
 import styled from 'styled-components';
 
@@ -137,6 +138,15 @@ function Component ({ className = '', currentTokenInfo, id, onCancel, tokenBalan
     });
   }, [balanceMap, currentAccount?.address, currentTokenInfo?.slug, isAllAccount]);
 
+  const symbol = currentTokenInfo?.symbol || '';
+
+  const filteredItems = useMemo(() => {
+    return accountItems
+      .filter((item) => {
+        return (new BigN(item.free).plus(item.locked)).gt(0);
+      });
+  }, [accountItems]);
+
   useEffect(() => {
     if (!isActive) {
       form?.resetFields();
@@ -194,12 +204,38 @@ function Component ({ className = '', currentTokenInfo, id, onCancel, tokenBalan
         {
           view === ViewValue.DETAIL && (
             <>
-              {accountItems.map((item) => (
-                <AccountTokenBalanceItem
-                  item={item}
-                  key={item.address}
-                />
-              ))}
+              {filteredItems.length
+                ? (filteredItems.map((item) => (
+                  <AccountTokenBalanceItem
+                    item={item}
+                    key={item.address}
+                  />
+                )))
+                : (
+                  <>
+                    <EmptyList
+                      buttonProps={{
+                        icon: <Icon
+                          phosphorIcon={ArrowCircleLeft}
+                          weight={'fill'}
+                        />,
+                        onClick: onCancel,
+                        size: 'xs',
+                        shape: 'circle',
+                        children: t('Back to home')
+                      }}
+                      className='__empty-list'
+                      emptyMessage={t('Switch to another token to see account balance')}
+                      emptyTitle={t('No account with {{symbol}} balance found', {
+                        replace: {
+                          symbol: symbol
+                        }
+                      })}
+                      key='empty-list'
+                      phosphorIcon={Coins}
+                    />
+                  </>
+                )}
             </>
           )
         }