+++ packages/extension-koni-ui/src/components/BackgroundExpandView.tsx
@@ -29,6 +29,8 @@ const Component: React.FC<Props> = (props: Props) => {
   const isShowFooter = useMemo(() => {
     const pathName = location.pathname;
 
+    console.log(pathName);
+
     return !['/create-done'].includes(pathName);
   }, [location.pathname]);